+++ src/app/Coriolis.jsx
@@ -174,8 +174,8 @@ export default class Coriolis extends React.Component {
           this._showModal(<ModalImport />);
         }
         break;
-      case 76:  // 'l'
-        if (e.ctrlKey || e.metaKey) { // CTRL/CMD + l
+      case 79:  // 'o'
+        if (e.ctrlKey || e.metaKey) { // CTRL/CMD + o
           e.preventDefault();
           this._showModal(<ModalPermalink url={window.location.href}/>);
         }