+++ ChangeLog.md
@@ -2,6 +2,7 @@
   * Tidy up old references to coriolis.io
   * Add ability to add and remove special effects to weapon modifications
   * Add weapon engineering information to Damage Dealt section
+  * Change shortcut for link from ctrl-l to ctrl-o to avoid clash with location bar
 
 #2.2.11
   * Add help system and initial help file